moved remaining traits to mod.rs, removed input.rs
